+++ lib/wavefront-cli/base.rb
@@ -295,7 +295,7 @@ module WavefrontCli
     def parseable_output(output_format, resp)
       options[:class] = klass_word
       options[:hcl_fields] = hcl_fields
-      cli_output_class.new(resp, options).run
+      cli_output_class(output_format).new(resp, options).run
     rescue LoadError
       raise(WavefrontCli::Exception::UnsupportedOutput,
             unsupported_format_message(output_format))